+++ tests/TwitterImageTest.php
@@ -7,7 +7,7 @@ use NotificationChannels\Twitter\TwitterImage;
 class TwitterImageTest extends TestCase
 {
     /** @test */
-    public function it_accepts_an_image_path_when_constructing_a_twitter_image()
+    public function it_accepts_an_image_path_when_constructing_a_twitter_image(): void
     {
         $image = new TwitterImage('/foo/bar/baz.png');
         $this->assertEquals('/foo/bar/baz.png', $image->getPath());